test(consensus_state): move state_monitor to common_test.h
